+++ DbgHelpUtils/print_utils.h
@@ -60,9 +60,9 @@ namespace dlg_help_utils::print_utils
     }
 
     template<typename T>
-    void print_stream_str(std::wostream& os, mini_dump_memory_stream& stream, size_t const max_size, bool const stop_at_null)
+    void print_stream_str(std::wostream& os, mini_dump_memory_stream& stream, uint64_t const max_size, bool const stop_at_null)
     {
-        for(size_t i = 0; i < max_size; ++i)
+        for(uint64_t i = 0; i < max_size; ++i)
         {
             T ch;
             if(stream.read(&ch, sizeof T) != sizeof T)
@@ -127,12 +127,12 @@ namespace dlg_help_utils::print_utils
     }
 
     template<typename T>
-    void print_stream_array_str(std::wostream& os, mini_dump_memory_stream& stream, size_t const max_size, size_t const indent = 0)
+    void print_stream_array_str(std::wostream& os, mini_dump_memory_stream& stream, uint64_t const max_size, size_t const indent = 0)
     {
         std::wstring const indent_str(indent, L' ');
         auto null_found = false;
         auto first = true;
-        for(size_t i = 0; i < max_size; ++i)
+        for(uint64_t i = 0; i < max_size; ++i)
         {
             T ch;
             if(stream.read(&ch, sizeof T) != sizeof T)
@@ -291,13 +291,13 @@ namespace dlg_help_utils::print_utils
     }
 
     template<typename T>
-    void print_stream_array_lines(std::wostream& os, mini_dump_memory_stream& stream, size_t const max_size, size_t const elements_per_line, size_t const element_width, size_t const indent, bool const dump_hex = false)
+    void print_stream_array_lines(std::wostream& os, mini_dump_memory_stream& stream, uint64_t const max_size, size_t const elements_per_line, size_t const element_width, size_t const indent, bool const dump_hex = false)
     {
         std::wstring const indent_str(indent, ' ');
 
         auto const index_width = find_max_hex_width(max_size - 1);
         auto first = true;
-        for(size_t i = 0; i < max_size; ++i)
+        for(uint64_t i = 0; i < max_size; ++i)
         {
             T value;
             if(stream.read(&value, sizeof T) != sizeof T)
